Fixed Some Bugs, Added Actuator UI

+++ src/main/java/com/teammusika/musika/services/RatingRepositoryService.java
@@ -28,21 +28,23 @@ public class RatingRepositoryService {
     public void storeRating(User user, Song song) {
     	List<Rating> ratings = findAll();
     	for (Rating rat : ratings){
-    		Set<Song> songs = rat.getSongs();
-    		for (Song songItem : songs) {
-    			if(song.getSongId()==songItem.getSongId()) {
-    				if(rat.getUser().getId()==user.getId()) {
-    					song.subSongLikes();
-    			        ratingRepository.delete(rat);
-    					songRepository.save(song);
-    					return;
+    		if(rat.getUser().getId()==user.getId()) {
+    			if(rat.getSongs().contains(song)) {
+        			if(rat.getSongs().size()==1) ratingRepository.delete(rat);
+        			else {
+        				rat.getSongs().remove(song);
+        				ratingRepository.save(rat);
     				}
-    				else {
-    					song.addSongLikes();
-    					songRepository.save(song);
-    					ratingRepository.save(rat);
-    					return;
-    				}
+    				song.subSongLikes();
+    				songRepository.save(song);
+    				return;
+    			}
+    			else {
+    				song.addSongLikes();
+    				songRepository.save(song);
+    				rat.addSongs(song);
+    		        ratingRepository.save(rat);
+    		        return;
     			}
     		}
     	}
@@ -52,7 +54,36 @@ public class RatingRepositoryService {
 		rating.setUser(user);
 		rating.addSongs(song);
         ratingRepository.save(rating);
-		return;
+    	return;
+//    		Set<Song> songs = rat.getSongs();
+//    		for (Song songItem : songs) {
+//    			if(song.getSongId()==songItem.getSongId()) {
+//    				if(rat.getUser().getId()==user.getId()) {
+//    			        ratingRepository.delete(rat);
+//    					song.subSongLikes();
+//    					songRepository.save(song);
+//    					return;
+//    				}
+//    				else {
+//    					song.addSongLikes();
+//    					songRepository.save(song);
+//    					Rating rating = new Rating();
+//    					rating.setUser(user);
+//    					rating.addSongs(song);
+//    			        ratingRepository.save(rating);
+////    					ratingRepository.save(rat);
+//    					return;
+//    				}
+//    			}
+//    		}
+//    	}
+//		song.addSongLikes();
+//		songRepository.save(song);
+//		Rating rating = new Rating();
+//		rating.setUser(user);
+//		rating.addSongs(song);
+//        ratingRepository.save(rating);
+//		return;
     }
     public List<Rating> findAll(){
         return ratingRepository.findAll();
